+++ src/main/java/com/phnanhky/matrix/entity/Question.java
@@ -20,7 +20,7 @@ import java.util.List;
 public class Question {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer questionId;
+    private Long questionId;
 
     @ManyToOne
     @JoinColumn(name = "lessonId")
